Added creator tools and fixed simfphys wheels + blinkers
https://youtu.be/8gcsB51C6lM

Simfphys vehicles now use blinkers when alarm is enabled
Simfphys vehicles that are locked can no longer be entered when clicking E on the wheels
Using a looped wav sound instead of looping the sound with timers (lights are still the same)

Given vehicle creators the ability to add support to their vehicle without the need of using the config
self:SetNWBool("carkeysSupported", true)
Bypasses the existing check for the class list, only problem is the vehicle isnt given to the player that spawned it. (only applies to vehicles with this bool, normal config added vehicles remain the same. Players just need to acquire the vehicle with a key after spawning)

Abliity to add custom alarms for vehicles.
self:SetNWBool("carkeysCustomAlarm", true)
self:SetNWString("carkeysCAlarmSound", "soundtablename") (soundtablename is the name of the sound when made with sound.add)

And last but probably one of the more important ones..
	self:SetNWFloat("carkeysForwardPos", 200 )
	self:SetNWFloat("carkeysRightPos", 0 )
	self:SetNWFloat("carkeysUpPos", 0 )
The code that finds how far you are from the vehicle gets the middle of the vehicle as the vehicles position, making larger vehicles like limos or planes not possible to unlock without being 150 units of the absolute center of the model.
These NW Floats allow you to change the centers position so vehicles like limos are possible to unlock near the actual doors.

Thats all that was changed.